pyproject.toml
[tool.poetry]
name = "pushupclub"
version = "0.1.0"
description = ""
authors = ["kuchev.sergey <kuchevwork@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.6"
aiogram = "2.25.1"
APScheduler = "3.10.4"
google_api_python_client = "2.97.0"
google_auth_oauthlib = "1.0.0"
protobuf = "4.24.2"
psycopg2 = "2.9.5"
psycopg2-binary = "2.9.5"
pytz = "2023.3"
uvicorn = "0.22.0"
gunicorn = "^20.1.0"
multidict = "6.0.4"


[tool.poetry.dev-dependencies]
devtools = "^0.11.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
